python/agents/developer-productivity-agent/pyproject.toml
[project]
name = "developer_productivity_agent"
version = "0.1.0"
description = "AI Agent for app developers to get information from Github, Stack Overflow and Confluence."
authors = [
    {name = "Vidya Ravindran",email = "ravindranv@google.com"}
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk (>=1.5.0,<2.0.0)",
    "google-cloud-aiplatform[adk,agent-engines] (>=1.100.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-cloud-secret-manager (>=2.24.0,<3.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
google-adk = ">=1.5.0,<2.0.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.2"
google-cloud-aiplatform = { version = ">=1.100.0", extras = ["adk","agent-engines"] }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
